requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.5.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pathspec==0.7.0
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
python-dotenv==0.12.0
pytz==2019.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
spotipy==2.10.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
